Refactored data access files in signup use_case to reduce redundancy
