+++ src/main/Java/use_case/signup/UserSignupDataAccessInterface.java
@@ -1,4 +1,4 @@
-package use_case.signup.general_user_signup;
+package use_case.signup;
 
 import entity.Account;
