--- src/main/Java/use_case/signup/choose_account_type/AccountTypeUserDataAccessInterface.java
@@ -1,4 +1,0 @@
-package use_case.signup.choose_account_type;
-
-public interface AccountTypeUserDataAccessInterface {
-}
