--- src/main/Java/use_case/signup/event_poster_signup/EventPosterUserDataAccessInterface.java
@@ -1,18 +1,0 @@
-package use_case.signup.event_poster_signup;
-
-import entity.Account;
-
-public interface EventPosterUserDataAccessInterface {
-    /**
-     * Checks if the given username exists.
-     * @param username the username to look for
-     * @return true if a user with the given username exists; false otherwise
-     */
-    boolean existsByName(String username);
-
-    /**
-     * Saves the user.
-     * @param user the user to save
-     */
-    void save(Account user);
-}
